chore: rename functions and files
